Move traits and types enums to their own modules.
