+++ google-smart-home/src/device/types.rs
@@ -1,0 +1,16 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize, Hash)]
+#[serde(rename_all = "camelCase")]
+#[non_exhaustive]
+pub enum Type {
+    #[serde(rename = "action.devices.types.GARAGE")]
+    Garage,
+    #[serde(rename = "action.devices.types.GATE")]
+    Gate,
+    #[serde(rename = "action.devices.types.LIGHT")]
+    Light,
+    #[serde(rename = "action.devices.types.OUTLET")]
+    Outlet,
+}
